+++ src/app/dashboard/layout.tsx
@@ -5,8 +5,8 @@ import Container from "@mui/material/Container";
 import GlobalStyles from "@mui/material/GlobalStyles";
 
 import { AuthGuard } from "@/components/auth/AuthGuard";
-import { MainNav } from "@/components/dashboard/layout/MainNav";
-import { SideNav } from "@/components/dashboard/layout/SideNav";
+import { MainNav } from "@/components/layout/MainNav";
+import { SideNav } from "@/components/layout/SideNav";
 import { config } from "@/config";
 
 export const metadata = {
