+++ src/components/item/schema.ts
