+++ src/components/supplier/CreateSupplierModal.tsx
@@ -12,11 +12,7 @@ import InputLabel from "@mui/material/InputLabel";
 import OutlinedInput from "@mui/material/OutlinedInput";
 import FormHelperText from "@mui/material/FormHelperText";
 import Stack from "@mui/material/Stack";
-import {
-  schema,
-  defaultValues,
-  Values,
-} from "@/components/dashboard/supplier/schema";
+import { schema, defaultValues, Values } from "@/components/supplier/schema";
 
 type CreateSupplierModalProps = {
   open: boolean;
